+++ examples/psql/patch/patch_operations.tf
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, 2023, Oracle and/or its affiliates. All rights reserved.
+// Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
 
 # Vars provided when configuring tf for oci
 variable "tenancy_ocid" {