+++ internal/resourcediscovery/export_compartment.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, 2023, Oracle and/or its affiliates. All rights reserved.
+// Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
 // Licensed under the Mozilla Public License v2.0
 
 package resourcediscovery
@@ -18,10 +18,9 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-multierror"
+	"github.com/hashicorp/go-version"
 
 	tf_export "github.com/oracle/terraform-provider-oci/internal/commonexport"
-
-	"github.com/hashicorp/go-version"
 
 	"github.com/oracle/terraform-provider-oci/internal/globalvar"
 	"github.com/oracle/terraform-provider-oci/internal/tfresource"