+++ internal/service/generative_ai/generative_ai_endpoints_data_source.go
@@ -1,0 +1,144 @@
+// Copyright (c) 2017, 2023, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+package generative_ai
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	oci_generative_ai "github.com/oracle/oci-go-sdk/v65/generativeai"
+
+	"github.com/oracle/terraform-provider-oci/internal/client"
+	"github.com/oracle/terraform-provider-oci/internal/tfresource"
+)
+
+func GenerativeAiEndpointsDataSource() *schema.Resource {
+	return &schema.Resource{
+		Read: readGenerativeAiEndpoints,
+		Schema: map[string]*schema.Schema{
+			"filter": tfresource.DataSourceFiltersSchema(),
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"id": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"state": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"endpoint_collection": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+
+						"items": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Elem:     tfresource.GetDataSourceItemSchema(GenerativeAiEndpointResource()),
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func readGenerativeAiEndpoints(d *schema.ResourceData, m interface{}) error {
+	sync := &GenerativeAiEndpointsDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*client.OracleClients).GenerativeAiClient()
+
+	return tfresource.ReadResource(sync)
+}
+
+type GenerativeAiEndpointsDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_generative_ai.GenerativeAiClient
+	Res    *oci_generative_ai.ListEndpointsResponse
+}
+
+func (s *GenerativeAiEndpointsDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *GenerativeAiEndpointsDataSourceCrud) Get() error {
+	request := oci_generative_ai.ListEndpointsRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if id, ok := s.D.GetOkExists("id"); ok {
+		tmp := id.(string)
+		request.Id = &tmp
+	}
+
+	if state, ok := s.D.GetOkExists("state"); ok {
+		request.LifecycleState = oci_generative_ai.EndpointLifecycleStateEnum(state.(string))
+	}
+
+	request.RequestMetadata.RetryPolicy = tfresource.GetRetryPolicy(false, "generative_ai")
+
+	response, err := s.Client.ListEndpoints(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	request.Page = s.Res.OpcNextPage
+
+	for request.Page != nil {
+		listResponse, err := s.Client.ListEndpoints(context.Background(), request)
+		if err != nil {
+			return err
+		}
+
+		s.Res.Items = append(s.Res.Items, listResponse.Items...)
+		request.Page = listResponse.OpcNextPage
+	}
+
+	return nil
+}
+
+func (s *GenerativeAiEndpointsDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(tfresource.GenerateDataSourceHashID("GenerativeAiEndpointsDataSource-", GenerativeAiEndpointsDataSource(), s.D))
+	resources := []map[string]interface{}{}
+	endpoint := map[string]interface{}{}
+
+	items := []interface{}{}
+	for _, item := range s.Res.Items {
+		items = append(items, EndpointSummaryToMap(item))
+	}
+	endpoint["items"] = items
+
+	if f, fOk := s.D.GetOkExists("filter"); fOk {
+		items = tfresource.ApplyFiltersInCollection(f.(*schema.Set), items, GenerativeAiEndpointsDataSource().Schema["endpoint_collection"].Elem.(*schema.Resource).Schema)
+		endpoint["items"] = items
+	}
+
+	resources = append(resources, endpoint)
+	if err := s.D.Set("endpoint_collection", resources); err != nil {
+		return err
+	}
+
+	return nil
+}