+++ internal/service/stack_monitoring/stack_monitoring_process_sets_data_source.go
@@ -1,0 +1,127 @@
+// Copyright (c) 2017, 2023, Oracle and/or its affiliates. All rights reserved.
+// Licensed under the Mozilla Public License v2.0
+
+package stack_monitoring
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	oci_stack_monitoring "github.com/oracle/oci-go-sdk/v65/stackmonitoring"
+
+	"github.com/oracle/terraform-provider-oci/internal/client"
+	"github.com/oracle/terraform-provider-oci/internal/tfresource"
+)
+
+func StackMonitoringProcessSetsDataSource() *schema.Resource {
+	return &schema.Resource{
+		Read: readStackMonitoringProcessSets,
+		Schema: map[string]*schema.Schema{
+			"filter": tfresource.DataSourceFiltersSchema(),
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"process_set_collection": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+
+						"items": {
+							Type:     schema.TypeList,
+							Computed: true,
+							Elem:     tfresource.GetDataSourceItemSchema(StackMonitoringProcessSetResource()),
+						},
+					},
+				},
+			},
+		},
+	}
+}
+
+func readStackMonitoringProcessSets(d *schema.ResourceData, m interface{}) error {
+	sync := &StackMonitoringProcessSetsDataSourceCrud{}
+	sync.D = d
+	sync.Client = m.(*client.OracleClients).StackMonitoringClient()
+
+	return tfresource.ReadResource(sync)
+}
+
+type StackMonitoringProcessSetsDataSourceCrud struct {
+	D      *schema.ResourceData
+	Client *oci_stack_monitoring.StackMonitoringClient
+	Res    *oci_stack_monitoring.ListProcessSetsResponse
+}
+
+func (s *StackMonitoringProcessSetsDataSourceCrud) VoidState() {
+	s.D.SetId("")
+}
+
+func (s *StackMonitoringProcessSetsDataSourceCrud) Get() error {
+	request := oci_stack_monitoring.ListProcessSetsRequest{}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	request.RequestMetadata.RetryPolicy = tfresource.GetRetryPolicy(false, "stack_monitoring")
+
+	response, err := s.Client.ListProcessSets(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response
+	request.Page = s.Res.OpcNextPage
+
+	for request.Page != nil {
+		listResponse, err := s.Client.ListProcessSets(context.Background(), request)
+		if err != nil {
+			return err
+		}
+
+		s.Res.Items = append(s.Res.Items, listResponse.Items...)
+		request.Page = listResponse.OpcNextPage
+	}
+
+	return nil
+}
+
+func (s *StackMonitoringProcessSetsDataSourceCrud) SetData() error {
+	if s.Res == nil {
+		return nil
+	}
+
+	s.D.SetId(tfresource.GenerateDataSourceHashID("StackMonitoringProcessSetsDataSource-", StackMonitoringProcessSetsDataSource(), s.D))
+	resources := []map[string]interface{}{}
+	processSet := map[string]interface{}{}
+
+	items := []interface{}{}
+	for _, item := range s.Res.Items {
+		items = append(items, ProcessSetSummaryToMap(item))
+	}
+	processSet["items"] = items
+
+	if f, fOk := s.D.GetOkExists("filter"); fOk {
+		items = tfresource.ApplyFiltersInCollection(f.(*schema.Set), items, StackMonitoringProcessSetsDataSource().Schema["process_set_collection"].Elem.(*schema.Resource).Schema)
+		processSet["items"] = items
+	}
+
+	resources = append(resources, processSet)
+	if err := s.D.Set("process_set_collection", resources); err != nil {
+		return err
+	}
+
+	return nil
+}