+++ scripts/gofmtcheck.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# Copyright (c) 2017, 2023, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
 
 # Check gofmt
 echo "==> Checking that code complies with gofmt requirements..."